move sql driver to storage dir
