+++ src/accounts/account.ts
@@ -1,7 +1,7 @@
 import { type Buffer } from 'node:buffer'
 import { randomBytes } from 'node:crypto'
 import { DEFAULT_PDATA_BASELINE, TOKEN_EXPIRATION_TIME } from '~constants.js'
-import { db } from '~db.js'
+import { db } from '~storage/sql.js'
 import { type OnlyProperties } from '~utils.js'
 
 // #region Account Class
